requirements.txt
rich>=13.0.0
Pillow>=9.0.0